LPS-77353 Split display context in two
