+++ modules/apps/web-experience/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/display/context/SiteAdminDisplayContext.java
@@ -20,7 +20,6 @@ import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.GroupConstants;
-import com.liferay.portal.kernel.model.LayoutSetPrototype;
 import com.liferay.portal.kernel.model.MembershipRequestConstants;
 import com.liferay.portal.kernel.model.OrganizationConstants;
 import com.liferay.portal.kernel.model.User;
@@ -34,7 +33,6 @@ import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
 import com.liferay.portal.kernel.service.GroupServiceUtil;
-import com.liferay.portal.kernel.service.LayoutSetPrototypeServiceUtil;
 import com.liferay.portal.kernel.service.MembershipRequestLocalServiceUtil;
 import com.liferay.portal.kernel.service.OrganizationLocalServiceUtil;
 import com.liferay.portal.kernel.service.UserGroupLocalServiceUtil;
@@ -42,25 +40,16 @@ import com.liferay.portal.kernel.service.UserLocalServiceUtil;
 import com.liferay.portal.kernel.service.permission.GroupPermissionUtil;
 import com.liferay.portal.kernel.service.permission.PortalPermissionUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
-import com.liferay.portal.model.impl.LayoutSetPrototypeImpl;
 import com.liferay.portal.service.persistence.constants.UserGroupFinderConstants;
 import com.liferay.portlet.usersadmin.search.GroupSearch;
-import com.liferay.site.admin.web.internal.constants.SiteAdminConstants;
 import com.liferay.site.admin.web.internal.constants.SiteAdminPortletKeys;
 import com.liferay.site.constants.SiteWebKeys;
-import com.liferay.site.util.GroupCreationStep;
-import com.liferay.site.util.GroupCreationStepRegistry;
 import com.liferay.site.util.GroupSearchProvider;
-import com.liferay.site.util.GroupStarterKit;
-import com.liferay.site.util.GroupStarterKitRegistry;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -69,7 +58,6 @@ import java.util.List;
 import javax.portlet.PortletURL;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
  * @author Pavel Savinov
@@ -78,42 +66,17 @@ import javax.servlet.http.HttpServletResponse;
 public class SiteAdminDisplayContext {
 
 	public SiteAdminDisplayContext(
-			HttpServletRequest request, HttpServletResponse response,
+			HttpServletRequest request,
 			LiferayPortletRequest liferayPortletRequest,
 			LiferayPortletResponse liferayPortletResponse)
 		throws Exception {
 
 		_request = request;
-		_response = response;
 		_liferayPortletRequest = liferayPortletRequest;
 		_liferayPortletResponse = liferayPortletResponse;
 
-		_groupCreationStepRegistry =
-			(GroupCreationStepRegistry)request.getAttribute(
-				SiteWebKeys.GROUP_CREATION_STEP_REGISTRY);
-
 		_groupSearchProvider = (GroupSearchProvider)request.getAttribute(
 			SiteWebKeys.GROUP_SEARCH_PROVIDER);
-
-		_groupStarterKitRegistry =
-			(GroupStarterKitRegistry)request.getAttribute(
-				SiteWebKeys.GROUP_STARTER_KIT_REGISTRY);
-
-		String creationStepName = ParamUtil.getString(
-			liferayPortletRequest, "creationStepName");
-
-		GroupCreationStep groupCreationStep =
-			_groupCreationStepRegistry.getGroupCreationStep(creationStepName);
-
-		if (groupCreationStep == null) {
-			List<GroupCreationStep> groupCreationSteps =
-				_groupCreationStepRegistry.getGroupCreationSteps(
-					request, response);
-
-			groupCreationStep = groupCreationSteps.get(0);
-		}
-
-		_groupCreationStep = groupCreationStep;
 	}
 
 	public int getChildSitesCount(Group group) {
@@ -122,68 +85,6 @@ public class SiteAdminDisplayContext {
 
 		return GroupLocalServiceUtil.getGroupsCount(
 			themeDisplay.getCompanyId(), group.getGroupId(), true);
-	}
-
-	public String getCreationStepRedirect() throws Exception {
-		PortletURL portletURL = _liferayPortletResponse.createRenderURL();
-
-		String groupCreationStepName = getNextCreationStepName();
-
-		if (Validator.isNull(groupCreationStepName)) {
-			portletURL.setParameter("jspPage", "/summary.jsp");
-
-			String redirect = ParamUtil.getString(_request, "redirect");
-
-			if (Validator.isNotNull(redirect)) {
-				portletURL.setParameter("redirect", redirect);
-			}
-
-			String backURL = ParamUtil.getString(_request, "backURL");
-
-			if (Validator.isNotNull(redirect)) {
-				portletURL.setParameter("backURL", backURL);
-			}
-		}
-		else {
-			portletURL.setParameter("jspPage", "/site_creation_wizard.jsp");
-
-			String redirect = ParamUtil.getString(_request, "redirect");
-
-			if (Validator.isNotNull(redirect)) {
-				portletURL.setParameter("redirect", redirect);
-			}
-
-			String creationStepName = ParamUtil.getString(
-				_request, "creationStepName");
-
-			if (Validator.isNotNull(creationStepName)) {
-				portletURL.setParameter("creationStepName", creationStepName);
-			}
-
-			String creationType = ParamUtil.getString(_request, "creationType");
-
-			if (Validator.isNotNull(creationStepName)) {
-				portletURL.setParameter("creationType", creationType);
-			}
-
-			String groupStarterKitKey = ParamUtil.getString(
-				_request, "groupStarterKitKey");
-
-			if (Validator.isNotNull(groupStarterKitKey)) {
-				portletURL.setParameter(
-					"groupStarterKitKey", groupStarterKitKey);
-			}
-		}
-
-		return portletURL.toString();
-	}
-
-	public String getCurrentCreationStepLabel() {
-		return _groupCreationStep.getLabel(_request);
-	}
-
-	public String getCurrentCreationStepName() {
-		return _groupCreationStep.getName();
 	}
 
 	public PortletURL getCurrentURL() {
@@ -219,11 +120,6 @@ public class SiteAdminDisplayContext {
 		return _group;
 	}
 
-	public List<GroupCreationStep> getGroupCreationSteps() throws Exception {
-		return _groupCreationStepRegistry.getGroupCreationSteps(
-			_request, _response);
-	}
-
 	public long getGroupId() {
 		if (_groupId <= 0) {
 			_groupId = ParamUtil.getLong(
@@ -233,78 +129,12 @@ public class SiteAdminDisplayContext {
 		return _groupId;
 	}
 
-	public String getGroupStarterKitDescription() {
-		String groupStarterKitKey = ParamUtil.getString(
-			_request, "groupStarterKitKey");
-
-		GroupStarterKit groupStarterKit =
-			_groupStarterKitRegistry.getGroupStarterKit(groupStarterKitKey);
-
-		if (groupStarterKit == null) {
-			return StringPool.BLANK;
-		}
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)_request.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		return groupStarterKit.getDescription(themeDisplay.getLocale());
-	}
-
-	public List<GroupStarterKit> getGroupStarterKits() {
-		return _groupStarterKitRegistry.getGroupStarterKits(
-			_group.getCompanyId(), true);
-	}
-
-	public String getGroupStarterKitThumbnailSrc() {
-		String groupStarterKitKey = ParamUtil.getString(
-			_request, "groupStarterKitKey");
-
-		GroupStarterKit groupStarterKit =
-			_groupStarterKitRegistry.getGroupStarterKit(groupStarterKitKey);
-
-		if (groupStarterKit == null) {
-			return StringPool.BLANK;
-		}
-
-		return groupStarterKit.getThumbnailSrc();
-	}
-
 	public String getKeywords() {
 		if (_keywords == null) {
 			_keywords = ParamUtil.getString(_request, "keywords");
 		}
 
 		return _keywords;
-	}
-
-	public List<LayoutSetPrototype> getLayoutSetPrototypes()
-		throws PortalException {
-
-		List<LayoutSetPrototype> layoutSetPrototypes = new ArrayList<>();
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)_request.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		LayoutSetPrototype defaultLayoutSetPrototype =
-			new LayoutSetPrototypeImpl();
-
-		defaultLayoutSetPrototype.setActive(true);
-		defaultLayoutSetPrototype.setCompanyId(themeDisplay.getCompanyId());
-		defaultLayoutSetPrototype.setLayoutSetPrototypeId(0L);
-
-		String blankSiteLayoutSetPrototypeName = LanguageUtil.get(
-			_request, "blank-site");
-
-		defaultLayoutSetPrototype.setName(
-			blankSiteLayoutSetPrototypeName, themeDisplay.getLocale());
-
-		layoutSetPrototypes.add(defaultLayoutSetPrototype);
-
-		layoutSetPrototypes.addAll(
-			LayoutSetPrototypeServiceUtil.search(
-				themeDisplay.getCompanyId(), Boolean.TRUE, null));
-
-		return layoutSetPrototypes;
 	}
 
 	public List<NavigationItem> getNavigationItems() {
@@ -371,40 +201,6 @@ public class SiteAdminDisplayContext {
 		return portletURL;
 	}
 
-	public String getPreviousCreationStepName() throws Exception {
-		GroupCreationStep groupCreationStep =
-			_groupCreationStepRegistry.getPreviousGroupCreationStep(
-				_groupCreationStep.getName(), _request, _response);
-
-		if (groupCreationStep == null) {
-			return null;
-		}
-
-		return groupCreationStep.getName();
-	}
-
-	public String getRenderPreviewURL(
-		String functionName, String title, String url) {
-
-		StringBundler sb = new StringBundler(13);
-
-		sb.append("javascript:");
-		sb.append(_liferayPortletResponse.getNamespace());
-		sb.append(functionName);
-		sb.append(StringPool.OPEN_PARENTHESIS);
-		sb.append(StringPool.APOSTROPHE);
-		sb.append(HtmlUtil.escapeJS(title));
-		sb.append(StringPool.APOSTROPHE);
-		sb.append(StringPool.COMMA_AND_SPACE);
-		sb.append(StringPool.APOSTROPHE);
-		sb.append(HtmlUtil.escapeJS(url));
-		sb.append(StringPool.APOSTROPHE);
-		sb.append(StringPool.CLOSE_PARENTHESIS);
-		sb.append(StringPool.SEMICOLON);
-
-		return sb.toString();
-	}
-
 	public GroupSearch getSearchContainer() throws PortalException {
 		return _groupSearchProvider.getGroupSearch(
 			_liferayPortletRequest, getPortletURL());
@@ -417,22 +213,6 @@ public class SiteAdminDisplayContext {
 		searchURL.setParameter("displayStyle", getDisplayStyle());
 
 		return searchURL;
-	}
-
-	public String getSummarySuccessMessage(String url) {
-		StringBundler sb = new StringBundler(3);
-
-		sb.append(
-			LanguageUtil.get(
-				_request, "congratulations-your-site-has-been-created"));
-		sb.append(StringPool.NEW_LINE);
-		sb.append(
-			LanguageUtil.format(
-				_request,
-				"you-can-finish-it-by-configuring-its-x-when-you-are-ready",
-				_getSummarySuccessMessageArgument(url)));
-
-		return sb.toString();
 	}
 
 	public int getUserGroupsCount(Group group) {
@@ -532,63 +312,13 @@ public class SiteAdminDisplayContext {
 		return false;
 	}
 
-	public boolean isLastGroupCreationStep() throws Exception {
-		String creationType = ParamUtil.getString(_request, "creationType");
-
-		if (!creationType.equals(
-				SiteAdminConstants.CREATION_TYPE_STARTER_KIT)) {
-
-			return false;
-		}
-
-		if (Validator.isNotNull(getNextCreationStepName())) {
-			return false;
-		}
-
-		return true;
-	}
-
-	public void renderCurrentCreationStep() throws Exception {
-		_groupCreationStep.render(_request, _response);
-	}
-
-	protected String getNextCreationStepName() throws Exception {
-		GroupCreationStep groupCreationStep =
-			_groupCreationStepRegistry.getNextGroupCreationStep(
-				_groupCreationStep.getName(), _request, _response);
-
-		if (groupCreationStep == null) {
-			return null;
-		}
-
-		return groupCreationStep.getName();
-	}
-
-	private String _getSummarySuccessMessageArgument(String url) {
-		StringBundler sb = new StringBundler(7);
-
-		sb.append("<a href=");
-		sb.append(StringPool.QUOTE);
-		sb.append(url);
-		sb.append(StringPool.QUOTE);
-		sb.append(StringPool.GREATER_THAN);
-		sb.append(LanguageUtil.get(_request, "site-settings"));
-		sb.append("</a>");
-
-		return sb.toString();
-	}
-
 	private String _displayStyle;
 	private Group _group;
-	private final GroupCreationStep _groupCreationStep;
-	private final GroupCreationStepRegistry _groupCreationStepRegistry;
 	private long _groupId;
 	private final GroupSearchProvider _groupSearchProvider;
-	private final GroupStarterKitRegistry _groupStarterKitRegistry;
 	private String _keywords;
 	private final LiferayPortletRequest _liferayPortletRequest;
 	private final LiferayPortletResponse _liferayPortletResponse;
 	private final HttpServletRequest _request;
-	private final HttpServletResponse _response;
 
 }
